requirements.txt
asgiref==3.4.1
Django==3.2.12
django-compressor==2.4.1
python-decouple==3.1
pytest-django==4.5.2
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2021.1
requests==2.24.0
sqlparse==0.4.2
